Add .prop v-bind support to vue-custom-elements.
By default, Vue allows the user to specify the method by which developers pass props (attribute or property). When encountering a custom element, Vue passes all props by attribute by default. This is not a good solution for rich data like objects and arrays. Vue, Angular, and other major frameworks provide a method to specify binding rich data to a component. For Vue, this is the '.prop' modifier.
https://vuejs.org/v2/api/#v-bind
The changes in this commit build in support for passing this rich data by checking for a passed-by-property prop if no matching attribute exists.